// src/api/types/ProxyResponse.ts
/**
 * This file was auto-generated by Fern from our API Definition.
 */

/**
 * The parsed response body from a proxied API request
 */
export interface ProxyResponse {
    /** HTTP status code */
    status?: number;
    /** Response headers */
    headers?: Record<string, unknown>;
    body?: unknown;
}